+++ src/view/InputView.js
@@ -1,4 +1,4 @@
-import Console from '../utils/Console';
+import Console from '../util/Console';
 
 const InputView = {
   CAR_NAMES_QUESTION: '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).',
